Extract PropertyForm and GenericForm into separate component files
